added dependencies
