+++ mobileapp/app.json
