+++ mobileapp/assets/adaptive-icon.png
