+++ mobileapp/package.json
@@ -9,13 +9,19 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@kamon/core": "1.0.0",
     "expo": "~48.0.15",
     "expo-status-bar": "~1.4.4",
     "react": "18.2.0",
     "react-native": "0.71.7"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0"
+    "@babel/core": "^7.20.0",
+    "@tsconfig/react-native": "^3.0.0",
+    "@types/jest": "^29.5.1",
+    "@types/react": "^18.2.0",
+    "@types/react-test-renderer": "^18.0.0",
+    "typescript": "^5.0.4"
   },
   "private": true
 }
