+++ webapp/package.json
@@ -24,7 +24,7 @@
     "db:migrate": "yarn run typeorm migration:run -d src/ormconfig.ts"
   },
   "dependencies": {
-    "@kamon/core": "workspace:^",
+    "@kamon/core": "1.0.0",
     "@nestjs/common": "^9.0.0",
     "@nestjs/config": "^2.3.1",
     "@nestjs/core": "^9.0.0",
